feat(worker): move all incident feed polling into the incidents worker
The app's prod no longer talks to any transit agency. Every incident
fetcher (BART, Sound Transit, NYC Subway, both Baltimore systems, CTA)
moves from src/lib/data.ts into the worker as-is, plus a new RTD Denver
fetcher that resolves alert entities through the system's id map with a
station-name text fallback. The worker serves generic
/incidents/:systemId and /incidents/:systemId/:stationId routes from KV,
refreshes everything on the hourly cron, and reports per-system health.

RTD's id map and station names and NYC's station name index are bundled
into the worker at deploy time, so the deploy workflow now also triggers
when those data files change.

The MTA Maryland feed is fetched over plain http on purpose: its TLS
config is incompatible with the workers runtime, and the public feed is
served identically on both.

getIncidents in the app is now a thin proxy over INCIDENTS_WORKER_URL
with the same 5 minute in-memory cache, and gtfs-realtime-bindings moves
out of the app's dependencies into the worker's.
