+++ package.json
@@ -26,7 +26,6 @@
     "csv-parse": "^6.2.1",
     "framer-motion": "^12.31.0",
     "fuse.js": "^7.1.0",
-    "gtfs-realtime-bindings": "^1.1.1",
     "jszip": "^3.10.1",
     "leaflet": "^1.9.4",
     "lucide-react": "^0.563.0",
